backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
agno==2.5.17
anthropic==0.96.0
openai==1.58.1
pandas==2.2.3
sqlalchemy==2.0.36
python-dateutil==2.9.0
requests==2.32.3
aiohttp==3.11.11
python-json-logger==3.2.1
httpx==0.28.1
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
slowapi==0.1.9
langsmith==0.1.68